Fixed DashedLine bug

+++ Fog Projekt/src/main/java/dat/backend/model/services/SVG.java
@@ -44,11 +44,11 @@ public class SVG {
     }
 
     public void addLine(int x1, int y1, int x2, int y2){
-        svgString.append(String.format(LINETEMPLATE, x1, y1, x2, y2));
+        svgString.append(String.format(LINETEMPLATE, (float)x1, (float)y1, (float)x2, (float)y2));
     }
 
     public void addDashedLine(int x1, int y1, int x2, int y2){
-        svgString.append(String.format(DASHEDLINETEMPLATE, x1, y1, x2, y2));
+        svgString.append(String.format(DASHEDLINETEMPLATE, (float)x1, (float)y1, (float)x2, (float)y2));
     }
 
     public void addInnerSvg(SVG innerSVGDrawing){
